Patching up files before starting on unit tests
