--- src/riff/new/iRiffChunk.ts
@@ -1,9 +1,0 @@
-import {ByteVector} from "../../byteVector";
-
-export default interface IRiffChunk {
-    chunkStart: number|undefined;
-    fourcc: string;
-    originalTotalSize: number;
-
-    render(): ByteVector;
-}
