+++ src/riff/riffFile.ts
@@ -1,40 +1,57 @@
+import AviHeader from "./avi/aviHeader";
 import CombinedTag from "../combinedTag";
 import DivxTag from "./divxTag";
 import Id3v2Tag from "../id3v2/id3v2Tag";
 import InfoTag from "./infoTag";
+import IRiffChunk from "./iRiffChunk";
 import MovieIdTag from "./movieIdTag";
 import Properties from "../properties";
+import RiffChunk from "./riffChunk";
+import RiffList from "./riffList";
 import RiffWaveFormatEx from "./riffWaveFormatEx";
-import {AviHeaderList} from "./aviHeaderList";
 import {ByteVector} from "../byteVector";
+import {CorruptFileError, UnsupportedFormatError} from "../errors";
 import {File, FileAccessMode, ReadStyle} from "../file";
 import {IFileAbstraction} from "../fileAbstraction";
-import {CorruptFileError, UnsupportedFormatError} from "../errors";
 import {ICodec} from "../iCodec";
 import {Id3v2TagHeaderFlags} from "../id3v2/id3v2TagHeader";
 import {Tag, TagTypes} from "../tag";
 import {NumberUtils} from "../utils";
 
 export default class RiffFile extends File {
-    /**
-     * Identifier used to recognize a RIFF file.
-     */
-    public static readonly FILE_IDENTIFIER = ByteVector.fromString("RIFF", undefined, undefined, true);
+    public static readonly fileIdentifier = ByteVector.fromString("RIFF", undefined, undefined, true);
 
-    private readonly _tag: CombinedTag = new CombinedTag();
+    private readonly _combinedTag: CombinedTag;
+    private _divxChunkIndex: number;
     private _divxTag: DivxTag;
+    private _fileType: string;
+    private _id3v2ChunkIndex: number;
     private _id3v2Tag: Id3v2Tag;
+    private _infoTagListIndex: number;
     private _infoTag: InfoTag;
-    private _properties: Properties;
+    private _movidIdListIndex: number;
     private _movieIdTag: MovieIdTag;
+    private _originalReadStyle: ReadStyle;
+    private _properties: Properties;
+    private _rawChunks: IRiffChunk[] = [];
+    private _riffSize: number;
 
+    /**
+     * Constructs and initializes a new instance of a RIFF file based on the provided file/file path.
+     * @param file File abstraction or path to a file to open as a RIFF file
+     * @param propertiesStyle How in-depth to read the properties of the file
+     */
     public constructor(file: IFileAbstraction|string, propertiesStyle: ReadStyle) {
         super(file);
 
-        this.read(true, propertiesStyle);
+        // Read the file
+        this.read(propertiesStyle);
+        this._combinedTag = new CombinedTag();
+        this._combinedTag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
+        this._originalReadStyle = propertiesStyle;
 
         // Make sure we have all the possible tags at our disposal
-        // @TODO: Creating *all* the tag types should probably be optional
+        // @TODO: Creating *all* the tag types sould probably be optional
         this.getTag(TagTypes.Id3v2, true);
         this.getTag(TagTypes.RiffInfo, true);
         this.getTag(TagTypes.MovieId, true);
@@ -44,10 +61,10 @@ export default class RiffFile extends File {
     // #region Properties
 
     /** @inheritDoc */
-    public get tag(): Tag { return this._tag; }
+    public get properties(): Properties { return this._properties; }
 
     /** @inheritDoc */
-    public get properties(): Properties { return this._properties; }
+    public get tag(): Tag { return this._combinedTag; }
 
     // #endregion
 
@@ -63,7 +80,7 @@ export default class RiffFile extends File {
                     this._id3v2Tag = Id3v2Tag.fromEmpty();
                     this._id3v2Tag.version = 4; // @TODO: Is only v2.4 supported or should we fallback to default?
                     this._id3v2Tag.flags |= Id3v2TagHeaderFlags.FooterPresent;
-                    this._tag.copyTo(this._id3v2Tag, true);
+                    this._combinedTag.copyTo(this._id3v2Tag, true);
                 }
 
                 tag = this._id3v2Tag;
@@ -72,7 +89,7 @@ export default class RiffFile extends File {
             case TagTypes.RiffInfo:
                 if (!this._infoTag) {
                     this._infoTag = InfoTag.fromEmpty();
-                    this._tag.copyTo(this._infoTag, true);
+                    this._combinedTag.copyTo(this._infoTag, true);
                 }
 
                 tag = this._infoTag;
@@ -81,7 +98,7 @@ export default class RiffFile extends File {
             case TagTypes.MovieId:
                 if (!this._movieIdTag) {
                     this._movieIdTag = MovieIdTag.fromEmpty();
-                    this._tag.copyTo(this._movieIdTag, true);
+                    this._combinedTag.copyTo(this._movieIdTag, true);
                 }
 
                 tag = this._movieIdTag;
@@ -90,14 +107,14 @@ export default class RiffFile extends File {
             case TagTypes.DivX:
                 if (!this._divxTag) {
                     this._divxTag = DivxTag.fromEmpty();
-                    this._tag.copyTo(this._divxTag, true);
+                    this._combinedTag.copyTo(this._divxTag, true);
                 }
 
                 tag = this._movieIdTag;
                 break;
         }
 
-        this._tag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
+        this._combinedTag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
         return tag;
     }
 
@@ -117,7 +134,7 @@ export default class RiffFile extends File {
             this._divxTag = undefined;
         }
 
-        this._tag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
+        this._combinedTag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
     }
 
     /** @inheritDoc */
@@ -126,272 +143,261 @@ export default class RiffFile extends File {
 
         this.mode = FileAccessMode.Write;
         try {
-            const data = ByteVector.empty();
+            // Determine the boundaries of the tagging chunks
+            const taggingChunkIndexes = [];
+            if (this._divxChunkIndex !== undefined) { taggingChunkIndexes.push(this._divxChunkIndex); }
+            if (this._id3v2ChunkIndex !== undefined) { taggingChunkIndexes.push(this._id3v2ChunkIndex); }
+            if (this._infoTagListIndex !== undefined) { taggingChunkIndexes.push(this._infoTagListIndex); }
+            if (this._movidIdListIndex !== undefined ) { taggingChunkIndexes.push(this._movidIdListIndex); }
+            taggingChunkIndexes.sort();
 
-            // Enclose the ID3v2 tag in an "id3 " item and embed it as the first tag.
+            // Render the tags we have
+            const renderedTags = [];
+            const replacedChunks: Array<{chunk: IRiffChunk, newTotalSize: number}> = [];
             if (this._id3v2Tag) {
-                const id3v2TagData = this._id3v2Tag.render();
-                if (id3v2TagData.length > 10) {
-                    if (id3v2TagData.length % 2 === 1) {
-                        id3v2TagData.addByte(0);
-                    }
-                    data.addByteVector(ByteVector.concatenate(
-                        ByteVector.fromString("id3 "),
-                        ByteVector.fromUInt(id3v2TagData.length, false),
-                        id3v2TagData
-                    ));
-                }
+                // @TODO: Allow chunk ID to be configurable
+                const id3v2TagBytes = this._id3v2Tag.render();
+                const id3v2TagChunk = RiffChunk.fromData("id3 ", id3v2TagBytes);
+                replacedChunks.push({ chunk: id3v2TagChunk, newTotalSize: id3v2TagBytes.length });
+                renderedTags.push(id3v2TagBytes);
             }
-
-            // Embed "INFO" tag as the second tag
             if (this._infoTag) {
-                data.addByteVector(this._infoTag.renderEnclosed());
-            }
+                const infoTagBytes = this._infoTag.render();
+                replacedChunks.push({ chunk: this._infoTag.list, newTotalSize: infoTagBytes.length });
+                renderedTags.push(infoTagBytes);
 
-            // Embed MovieID as the third tag
+            }
             if (this._movieIdTag) {
-                data.addByteVector(this._movieIdTag.renderEnclosed());
-            }
+                const movieIdBytes = this._movieIdTag.render();
+                replacedChunks.push({ chunk: this._movieIdTag.list, newTotalSize: movieIdBytes.length });
+                renderedTags.push(movieIdBytes);
 
-            // Embed the DivX tag in "IDVX" as the fourth tag
+            }
             if (this._divxTag) {
-                const divxTagData = this._divxTag.render();
-                // NOTE: No need to pad odd sized tags, b/c DivX is always 128 bytes
-                data.addByteVector(ByteVector.concatenate(
-                    ByteVector.fromString("IDVX"),
-                    ByteVector.fromUInt(divxTagData.length, false),
-                    divxTagData
-                ));
+                const divxTagBytes = this._divxTag.render();
+                const divxTagChunk = RiffChunk.fromData(DivxTag.CHUNK_FOURCC, divxTagBytes);
+                replacedChunks.push({ chunk: divxTagChunk, newTotalSize: divxTagBytes.length });
+                renderedTags.push(divxTagBytes);
             }
+            const renderedTagBytes = ByteVector.concatenate(... renderedTags);
 
-            // Reread the file to figure out where the RIFF tag blocks should go
-            // NOTE: I suppose we could just store this info in the class when we first read in the
-            //    file, but this *does* prevent problems if the file changes in between reading and
-            //    wring.
-            const readInfo = this.read(false, ReadStyle.None);
+            // Did the file originally have tags?
+            let taggingChunkStartIndex: number;
+            let taggingChunkStart: number;
+            let taggingChunkLength: number = 0;
+            if (taggingChunkIndexes.length > 0) {
+                // Determine if the chunks are contiguous
+                let taggingChunksAreContiguous = true;
+                for (let i = 1; i < taggingChunkIndexes.length && taggingChunksAreContiguous; i++) {
+                    const chunkIndexA = taggingChunkIndexes[i - 1];
+                    const chunkIndexB = taggingChunkIndexes[i];
+                    let chunkIndexC = chunkIndexA + 1;
+                    while (chunkIndexC < chunkIndexB) {
+                        // It's ok if we have A-JUNK-JUNK-JUNK-B, but A-X-B is not ok
+                        const chunkC = this._rawChunks[chunkIndexC];
+                        if (chunkC.fourcc !== "JUNK") {
+                            taggingChunksAreContiguous = false;
+                            break;
+                        }
 
-            // If tagging info cannot be found, place it at the end of the file
-            if (readInfo.tagStart < 12 || readInfo.tagEnd < readInfo.tagStart) {
-                readInfo.tagStart = this.length;
-                readInfo.tagEnd = this.length;
-            }
-
-            const originalTagLength = readInfo.tagEnd - readInfo.tagStart;
-
-            // If the tag isn't at the end of the file, try appending using padding to improve
-            // write time now or for subsequent writes.
-            if (readInfo.tagEnd !== this.length) {
-                let paddingSize = originalTagLength - data.length - 8;
-                if (paddingSize < 0) {
-                    paddingSize = 1024;
+                        chunkIndexC++;
+                    }
                 }
 
-                data.addByteVector(ByteVector.concatenate(
-                    ByteVector.fromString("JUNK"),
-                    ByteVector.fromUInt(paddingSize, false),
-                    ByteVector.fromSize(paddingSize)
-                ));
+                // Are the chunks contiguous?
+                if (taggingChunksAreContiguous) {
+                    // Awesome, the chunks are contiguous, we can update this in one big write
+                    // operation
+                    // Find any trailing JUNK chunks
+                    let lastChunkIndex = taggingChunkIndexes[taggingChunkIndexes.length - 1];
+                    let junkSize = 0;
+                    while (lastChunkIndex + 1 < this._rawChunks.length && this._rawChunks[lastChunkIndex + 1].fourcc === "JUNK") {
+                        lastChunkIndex++;
+                        junkSize += this._rawChunks[lastChunkIndex].originalTotalSize;
+                    }
+
+                    // Calculate tagging chunks start/length
+                    taggingChunkStartIndex = taggingChunkIndexes[0];
+                    taggingChunkStart = this._rawChunks[taggingChunkStartIndex].chunkStart;
+                    taggingChunkLength = this._rawChunks[lastChunkIndex].chunkStart
+                        + this._rawChunks[lastChunkIndex].originalTotalSize
+                        - taggingChunkStart;
+
+                    // Remove the chunks from the chunk list
+                    // NOTE: After this point, the chunk list and file contents have diverged (but
+                    //    it's ok b/c we will update the whole thing as part of writing the tags
+                    //    back out)
+                    this._rawChunks.splice(taggingChunkStartIndex, lastChunkIndex - taggingChunkIndexes[0]);
+                } else {
+                    // Crud, we can't update in one big write. Let's delete each tagging chunk,
+                    // then write them all in a giant block later. This process will be slow on
+                    // large files, but future saves should be improved.
+                    for (const chunkIndex of taggingChunkIndexes) {
+                        // Find any trailing JUNK chunks
+                        let lastChunkIndex = chunkIndex;
+                        while (
+                            lastChunkIndex + 1 < this._rawChunks.length &&
+                            this._rawChunks[lastChunkIndex + 1].fourcc === "JUNK"
+                        ) {
+                            lastChunkIndex++;
+                        }
+
+                        // Remove the tag chunk and any JUNK chunks, update start positions for
+                        // subsequent chunks
+                        const blockStart = this._rawChunks[chunkIndex].chunkStart;
+                        const blockLength = this._rawChunks[lastChunkIndex].chunkStart
+                            + this._rawChunks[lastChunkIndex].originalTotalSize
+                            - blockStart;
+                        this.removeBlock(blockStart, blockLength);
+                        this._rawChunks.splice(chunkIndex, lastChunkIndex - chunkIndex);
+                        this.updateChunkPositions(chunkIndex, -blockLength);
+                    }
+                }
             }
 
-            // Insert the tagging data
-            this.insert(data, readInfo.tagStart, originalTagLength);
-
-            // If the data size changed, and the tagging data is within the RIFF portion of the
-            // file, update the riff size
-            if (data.length - originalTagLength !== 0 && readInfo.tagStart <= readInfo.riffSize) {
-                const newRiffSize = ByteVector.fromUInt(readInfo.riffSize + data.length - originalTagLength, false);
-                this.insert(newRiffSize, 4, 4);
+            // If we don't know where to insert the tags, try to insert them before movi/data chunk,
+            // failing that, insert at end of the file.
+            if (taggingChunkStart === undefined) {
+                const moviChunkId = this._rawChunks.findIndex((c) => c.fourcc === "movi" || c.fourcc === "data");
+                if (moviChunkId >= 0) {
+                    taggingChunkStartIndex = moviChunkId;
+                    taggingChunkStart = this._rawChunks[taggingChunkStartIndex].chunkStart;
+                } else {
+                    taggingChunkStartIndex = this._rawChunks.length;
+                    taggingChunkStart = this.length;
+                }
             }
 
-            // Update the tag types on disk
-            this._tagTypesOnDisk = this.tagTypes;
+            // Determine new JUNK chunk size
+            let junkLength: number;
+            if (renderedTagBytes.length + 8 < taggingChunkLength) {
+                // New tags are smaller, so use the difference as the size of the JUNK chunk
+                junkLength = taggingChunkLength - renderedTagBytes.length + 8;
+            } else {
+                // New tags are same size or larger, so insert a JUNK chunk for future padding
+                // @TODO: Let padding size be configurable
+                junkLength = 1024;
+            }
+            const junkChunk = RiffChunk.fromData("JUNK", ByteVector.fromSize(junkLength));
+            const junkBytes = junkChunk.render();
+            renderedTagBytes.addByteVector(junkBytes);
+            replacedChunks.push({ chunk: junkChunk, newTotalSize: junkBytes.length });
+
+            // Write the big block to the file
+            this.insert(renderedTagBytes, taggingChunkStart, taggingChunkLength);
+
+            // Update the chunk list
+            this.updateChunkPositions(taggingChunkStartIndex, renderedTagBytes.length - taggingChunkLength);
+            this._rawChunks.splice(taggingChunkStartIndex, 0, ... replacedChunks.map((rc) => rc.chunk));
+            let filePosition = taggingChunkStart;
+            for (const replacedChunk of replacedChunks) {
+                replacedChunk.chunk.chunkStart = filePosition;
+                replacedChunk.chunk.originalTotalSize = replacedChunk.newTotalSize;
+                filePosition += replacedChunk.chunk.originalTotalSize;
+            }
         } finally {
             this.mode = FileAccessMode.Closed;
         }
     }
 
-    private read(readTags: boolean, style: ReadStyle): {riffSize: number, tagEnd: number, tagStart: number} {
+    private read(readStyle: ReadStyle): void {
         this.mode = FileAccessMode.Read;
 
         try {
-            // Make sure this is a RIFF file
-            if (!ByteVector.equal(this.readBlock(4), RiffFile.FILE_IDENTIFIER)) {
+            // Read the header of the file
+            if (!ByteVector.equal(this.readBlock(4), RiffFile.fileIdentifier)) {
                 throw new CorruptFileError("File does not begin with RIFF identifier");
             }
+            this._riffSize = this.readBlock(4).toUInt(false);
+            this._fileType = this.readBlock(4).toString();
 
-            // Initialize our loop state
-            const riffSize = this.readBlock(4).toUInt(false);
-            const streamFormat = this.readBlock(4).toString();
-            let tagStart = -1;
-            let tagEnd = -1;
-            let position = 12;
-            let size = 0;
-            let durationMilliseconds = 0;
-            let codecs: ICodec[] = [];
-
-            // Read until there are less than 8 bytes to read
-            do {
-                let tagFound = false;
-
-                // Check if the current position is an odd number and increment it so it is even.
-                // This is done when the previous chunk size was an odd number. If this is not
-                // done, the chunk being read after the odd chunk will not be read.
-                if (position > 12 && (position % 2) === 1) {
-                    position++;
-                }
-
-                this.seek(position);
+            // Read chunks until there are less than 8 bytes to read
+            let position = this.position;
+            while (position + 8 < this.length) {
                 const fourcc = this.readBlock(4).toString();
-                size = this.readBlock(4).toUInt(false);
-
-                switch (fourcc) {
-                    case "fmt ":
-                        // "fmt " is used by wave files to hold the WaveFormatEx structure
-                        if (style === ReadStyle.None || streamFormat !== "WAVE") {
-                            break;
-                        }
-
-                        this.seek(position + 8);
-                        codecs = [new RiffWaveFormatEx(this.readBlock(18), 0)];
-                        break;
-
-                    case "data":
-                        // "data" contains the audio data for wave files. It's contents represent
-                        // the invariant portion of the file and is used to determine the duration
-                        // of a file. It should always appear after "fmt ".
-                        if (streamFormat !== "WAVE") {
-                            break;
-                        }
-
-                        this._invariantStartPosition = position;
-                        this._invariantEndPosition = position + size;
-
-                        if (
-                            style === ReadStyle.None ||
-                            codecs.length !== 1 ||
-                            !(codecs[0] instanceof RiffWaveFormatEx)
-                        ) {
-                            break;
-                        }
-
-                        const durationSeconds = size / (<RiffWaveFormatEx> codecs[0]).averageBytesPerSecond;
-                        durationMilliseconds += durationSeconds * 1000;
-                        break;
-
-                    case "LIST":
-                        // Lists are used to store a variety of data collections. Read the type and
-                        // act on it.
-                        switch (this.readBlock(4).toString()) {
-                            case "hdrl":
-                                // "hdrl" is used by AVI files to hold a media header and
-                                // BitmapInfoHeader and WaveFormatEv structures.
-                                if (style === ReadStyle.None || streamFormat !== "AVI ") {
-                                    position += 8 + size;
-                                    continue;
-                                }
-
-                                const headerList = new AviHeaderList(this, position + 12, size - 4);
-                                durationMilliseconds = headerList.header.durationMilliseconds;
-                                codecs = headerList.codecs;
-                                break;
-
-                            case "INFO":
-                                // "INFO" is a tagging format handled by the InfoTag class
-                                if (readTags && !this._infoTag) {
-                                    this._infoTag = InfoTag.fromFile(this, position + 12, size - 4);
-                                }
-
-                                tagFound = true;
-                                break;
-
-                            case "MID ":
-                                // "MID " is a tagging format handled by the MovieIdTag class
-                                if (readTags && !this._movieIdTag) {
-                                    this._movieIdTag = MovieIdTag.fromFile(this, position + 12, size - 4);
-                                }
-
-                                tagFound = true;
-                                break;
-
-                            case "movi":
-                                // "movi" contains the media data for an AVI and its contents
-                                // represent the invariant portion of the file
-                                if (streamFormat !== "AVI ") {
-                                    break;
-                                }
-
-                                this._invariantStartPosition = position;
-                                this._invariantEndPosition = position + size;
-                                break;
-                        }
-                        break;
-
-                    case "id3 ":
-                    case "ID3 ":
-                    case "ID32":
-                        // "ID32" is a custom box for this format that contains an ID3v2 tag.
-                        // "ID3 " and "id3 " have become de facto standards.
-                        if (readTags && !this._id3v2Tag) {
-                            this._id3v2Tag = Id3v2Tag.fromFile(this, position + 8, style);
-                        }
-
-                        tagFound = true;
-                        break;
-
-                    case "IDVX":
-                        // "IDVX" is used by DivX and holds an ID3v1 style tag
-                        if (readTags && !this._divxTag) {
-                            this._divxTag = DivxTag.fromFile(this, position + 8);
-                        }
-
-                        tagFound = true;
-                        break;
-
-                    case "JUNK":
-                        // "JUNK" is a padding element that could be associated with tag data.
-                        if (tagEnd === position) {
-                            tagEnd = position + 8 + size;
-                        }
-                        break;
-                }
-
-                // Determine the region of the file that contains tags.
-                if (tagFound) {
-                    if (tagStart === -1) {
-                        tagStart = position;
-                        tagEnd = position + 8 + size;
-                    } else if (tagEnd === position) {
-                        tagEnd = position + 8 + size;
-                    }
-                }
-
-                // Move to the next item
-                position += 8 + size;
-            } while (position + 8 < this.length);
-
-            // If we're reading properties and one was found, throw an exception. Otherwise
-            // create the properties object
-            if (style !== ReadStyle.None) {
-                if (codecs.length === 0) {
-                    throw new UnsupportedFormatError("Unsupported RIFF type");
-                }
-
-                this._properties = new Properties(durationMilliseconds, codecs);
+                const chunk = fourcc === RiffList.identifierFourcc
+                    ? RiffList.fromFile(this, position)
+                    : RiffChunk.fromFile(this, fourcc, position);
+                position += chunk.originalTotalSize;
+                this._rawChunks.push(chunk);
             }
 
-            // If we're reading tags, update the combined tag
-            if (readTags) {
-                this._tag.setTags(this._id3v2Tag, this._infoTag, this._movieIdTag, this._divxTag);
+            // Process the properties of the file
+            let codecs: ICodec[];
+            let durationMilliseconds = 0;
+            if ((readStyle & ReadStyle.Average) !== 0) {
+                switch (this._fileType) {
+                    case "WAVE":
+                        // This is a wav file, search for fmt chunk
+                        const fmtChunk = this._rawChunks.find((c) => c.fourcc === RiffWaveFormatEx.CHUNK_FOURCC);
+                        if (!fmtChunk) {
+                            throw new CorruptFileError("WAV file is missing header chunk");
+                        }
+
+                        const waveHeader = new RiffWaveFormatEx((<RiffChunk> fmtChunk).data);
+                        codecs = [waveHeader];
+                        durationMilliseconds = waveHeader.durationMilliseconds;
+                        break;
+
+                    case "AVI ":
+                        // This is an AVI file, search for the hdrl list
+                        const aviHeaderList = this._rawChunks.find((c) => {
+                            return c.fourcc === RiffList.identifierFourcc
+                                && (<RiffList> c).type === AviHeader.headerListType;
+                        });
+                        const aviHeader = new AviHeader(<RiffList> aviHeaderList);
+                        codecs = aviHeader.codecs;
+                        durationMilliseconds = aviHeader.durationMilliseconds;
+                        break;
+
+                    default:
+                        throw new UnsupportedFormatError("Unsupported RIFF type");
+                }
             }
 
-            return {
-                riffSize: riffSize,
-                tagEnd: tagEnd,
-                tagStart: tagStart
-            };
+            this._properties = new Properties(durationMilliseconds, codecs);
+
+            // Process tags
+            // 1) DivX
+            this._divxChunkIndex = this._rawChunks.findIndex((c) => c.fourcc === DivxTag.CHUNK_FOURCC);
+            if (this._divxChunkIndex !== undefined) {
+                this._divxTag = DivxTag.fromData((<RiffChunk> this._rawChunks[this._divxChunkIndex]).data);
+            }
+
+            // 2) ID3v2
+            const id3v2ChunkFourcc = ["id3 ", "ID3 ", "ID32"];
+            this._id3v2ChunkIndex = this._rawChunks.findIndex((c) => id3v2ChunkFourcc.indexOf(c.fourcc) >= 0);
+            if (this._id3v2ChunkIndex !== undefined) {
+                // @TODO: Switch to fromFile and using chunk start/data length to allow lazy picture loading
+                this._id3v2Tag = Id3v2Tag.fromData((<RiffChunk> this._rawChunks[this._id3v2ChunkIndex]).data);
+            }
+
+            // 3) Info tag
+            this._infoTagListIndex = this._rawChunks.findIndex((c) => {
+                return c.fourcc === RiffList.identifierFourcc
+                    && (<RiffList> c).type === InfoTag.listType;
+            });
+            if (this._infoTagListIndex !== undefined) {
+                this._infoTag = InfoTag.fromList(<RiffList> this._rawChunks[this._infoTagListIndex]);
+            }
+
+            // 4) MovieID tag
+            this._movidIdListIndex = this._rawChunks.findIndex((c) => {
+                return c.fourcc === RiffList.identifierFourcc
+                    && (<RiffList> c).type === MovieIdTag.listType;
+            });
+            if (this._movidIdListIndex !== undefined) {
+                this._movieIdTag = MovieIdTag.fromList(<RiffList> this._rawChunks[this._movidIdListIndex]);
+            }
 
         } finally {
             this.mode = FileAccessMode.Closed;
+        }
+    }
+
+    private updateChunkPositions(chunkIndex: number, filePositionOffset: number): void {
+        for (let i = chunkIndex; i < this._rawChunks.length; i++) {
+            this._rawChunks[i].chunkStart += filePositionOffset;
         }
     }
