+++ src/riff/riffListTag.ts
@@ -1,7 +1,7 @@
 import RiffList from "./riffList";
-import {ByteVector, StringType} from "../../byteVector";
-import {Tag} from "../../tag";
-import {Guards} from "../../utils";
+import {ByteVector, StringType} from "../byteVector";
+import {Tag} from "../tag";
+import {Guards} from "../utils";
 
 /**
  * Abstract class that provides support for reading/writing tags in the RIFF list format.
